refactor(fronted): add correct implementation of Dialog and Table
Following the rules atomic design
