+++ frontend/src/app/(auth)/layout.tsx
@@ -8,7 +8,7 @@ const DashboardLayout = ({
       <div className="w-1/3 lg:w-[20rem]">
         <Sidebar />
       </div>
-      <div className="w-2/3 lg:flex-grow lg:pl-4 bg-slate-700">{children}</div>
+      <div className="w-2/3 lg:flex-grow pt-8 pr-8 lg:pl-8">{children}</div>
     </section>
   );
 };
