+++ frontend/src/components/organisms/FormAssignor.tsx
@@ -2,7 +2,7 @@ import { FormField } from "../molecules/FormField";
 
 export const FormAssignor = () => {
   return (
-    <div className="flex flex-col gap-y-4 w-full h-full p-4">
+    <div className="flex flex-col gap-y-4 w-full h-full">
       <FormField title="Nome" />
       <FormField title="CPF/CNPJ" />
       <FormField title="Email" />
